modularize code, slight reformat & refactoring, add shorthand parameter
allow search() to take an array instead of an object as shorthand for
the `field` option

change
`if (key[0] === '_' || !this.hasOwnProperty(key))`
to
`if (!this.hasOwnProperty(key) || key[0] === '_')`
to better support IDE static analysis

remove unused `fuzzyDiacritics` declaration
