+++ lib/regexFieldSearch.js
@@ -1,0 +1,8 @@
+const transformSearchFieldsInQuery = require('./transformSearchFieldsInQuery')
+
+
+module.exports = function regexFieldSearch (sm) {
+  return function (hook) {
+    transformSearchFieldsInQuery(hook.params.query, sm)
+  }
+}
